% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoMcl.R
\name{clustered}
\alias{clustered}
\title{Estimating interventional (in)direct effects for multiple mediators with clustered data}
\usage{
clustered(
  data,
  Sname,
  Wnames = NULL,
  Xnames,
  Aname,
  Mnames,
  Yname,
  Yfamily = "gaussian",
  cluster_opt_a = "FE.glm",
  cluster_opt_m = "FE.glm",
  cluster_opt_y = "FE.glm",
  interaction_fitm2 = c("AM.M1"),
  interaction_fity = c("AM.M1", "AM.M2", "Mint", "AMint"),
  learners_a = c("SL.glm"),
  learners_m = c("SL.glm"),
  learners_y = c("SL.glm"),
  num_folds = 1
)
}
\arguments{
\item{data}{a data frame containing the variables of treatment A, mediators M1 and M2, outcome Y, cluster indicators S, individual-level covariates X, and cluster-level covariates W (if there are any).}

\item{Sname}{a character for the name of variable indicating the cluster membership of each individual. For example, Sname may be the school identification number in a two-level data structure with students clustered in schools.}

\item{Wnames}{a character vector for the names of the cluster-level covariates.}

\item{Xnames}{a character vector for the names of the individual-level covariates.}

\item{Aname}{a character for the name of the treatment variable.}

\item{Mnames}{a character vector for the names of the mediators. Currently support two mediators.}

\item{Yname}{a character for the name of the outcome variable.}

\item{Yfamily}{the link function for estimating the conditional mean outcome in a generalized linear model.}

\item{cluster_opt_a}{a character specifying the option for estimating the model for the treatment with clusters.}

\item{cluster_opt_m}{a character specifying the option for estimating the model for the mediators with clusters.}

\item{cluster_opt_y}{a character specifying the option for estimating the model for the outcome with clusters.}

\item{interaction_fitm2}{a character vector specifying the interaction terms to include in estimating the model for the mediator M2.}

\item{interaction_fity}{a character vector specifying the interaction terms to include in estimating the model for the outcome Y.}

\item{learners_a}{only needed if "cluster_opt_a" is "FE.mlr" or "noncluster.mlr", which is a character (vector) specifying the estimation methods to use with the SuperLearner R package (i.e., the argument "SL.library = " of the function "SuperLearner()").}

\item{learners_m}{only needed if "cluster_opt_m" is "FE.mlr" or "noncluster.mlr", which is a character (vector) specifying the estimation methods to use with the SuperLearner R package.}

\item{learners_y}{only needed if "cluster_opt_y" is "FE.mlr" or "noncluster.mlr", which is a character (vector) specifying the estimation methods to use with the SuperLearner R package.}

\item{num_folds}{only needed if cross-fitted estimates of the models are used, which is an integer specifying the number of folds to use for the cross-fitting procedure.}
}
\value{
The function returns a list containing lists of estimates from the three estimators. Specifically, a list named "thetas_effs" contains the estimates from the multiply-robust estimator; a list named "regs_effs" contains the estimates from the regression-based estimator; a list named "rmpw_effs" contains the estimates from the weighting-based estimator.
}
\description{
Estimating interventional (in)direct effects for multiple mediators with clustered data
}
\details{
For the arguments "cluster_opt_a", "cluster_opt_m" and "cluster_opt_y", currently supported options include: "FE.glm", "FE.mlr", "noncluster.glm", "noncluster.mlr", "RE.glm", in which "FE" stands for the fixed-effects method with cluster indicators included as part of the covariates, "noncluster" stands for ignoring the clusters, "RE" stands for the random-effects method with clusters as random intercepts; additionally, "glm" stands for parametric estimation with linear terms of covariates, "mlr" stands for machine learning estimation with the "SuperLearner" R package.
}
\examples{
dataset <- read.csv("data/dataset.csv")
Sname <- "school"
Xnames <- colnames(dataset)[grep("^X", colnames(dataset))]
Wnames <- colnames(dataset)[grep("^W", colnames(dataset))]
Aname <- "A"
Mnames <- colnames(dataset)[grep("^M", colnames(dataset))]
Yname <- "Y"
# out <- clustered(data = dataset,
#                  Sname = Sname,
#                  Wnames = Wnames,
#                  Xnames = Xnames,
#                  Aname = Aname,
#                  Mnames = Mnames,
#                  Yname = Yname,
#                  Yfamily = "gaussian",
#                  cluster_opt_a = "FE.glm",
#                  cluster_opt_m = "FE.glm",
#                  cluster_opt_y = "FE.glm",
#                  interaction_fitm2 =  c("AM.M1"),
#                  interaction_fity = c("AM.M1", "AM.M2", "Mint", "AMint")
# )
# out

}
