% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMed.R
\name{ClusterMed}
\alias{ClusterMed}
\title{Estimating Causal Mediation Effects in Multiple-Mediator Analyses with Clustered Data}
\usage{
ClusterMed(
  data,
  Sname,
  Wnames,
  Xnames,
  Aname,
  Mnames,
  Yname,
  learners_a = c("SL.glm"),
  learners_m = c("SL.glm"),
  learners_y = c("SL.glm"),
  cluster_opt = "cwc.FE",
  Morder = "12",
  num_folds = 1
)
}
\arguments{
\item{data}{a \code{data.frame} containing the analysis variables.}

\item{Sname}{a character string of the name of the column in "data" that indicates the cluster membership of each individual (S).}

\item{Wnames}{a character vector of the names of the columns in "data" that correspond to cluster-level baseline (i.e., pre-treatment) covariates (W). "Wnames" may include the cluster size (i.e., number of individuals in each cluster).}

\item{Xnames}{a character vector of the names of the columns in "data" that correspond to individual-level baseline covariates (X).}

\item{Aname}{a character string of the name of the column in "data" that corresponds to a dummy coded treatment variable (A).}

\item{Mnames}{a character vector of the names of the columns in "data" that correspond to mediators (M).}

\item{Yname}{a character string of the name of the column in "data" that corresponds to an outcome variable (Y).}

\item{learners_a}{a character vector specifying estimation methods for the model of the treatment For "learners_a", "learners_m", and "learners_y", currently supported options include methods in the "SuperLearner" package (<https://cran.r-project.org/package=SuperLearner>), available via the "SuperLearner::listWrappers()".}

\item{learners_m}{a character vector specifying estimation methods for the models of the mediators.}

\item{learners_y}{a character vector specifying estimation methods for the model of the outcome.}

\item{cluster_opt}{a character string of the option of incorporating clusters in the nuisance model estimation. Currently supported options are "cwc.FE" (the default option) and "cwc". "cwc.FE": Control for cluster means  and perform cluster-mean centering for individual-level variables, and also control for dummy indicators of an individual's  cluster membership. "cwc": Control for cluster means  and perform cluster-mean centering for individual-level variables (while omitting the cluster dummies).}

\item{num_folds}{the number of folds to use for the cross-fitting procedure.}
}
\value{
A data frame containing the results (estimates and 0.95 confidence intervals) for each of the cluster-average and individual-average causal mediation effects (i.e., IDE, IIE_M1, IIE_M2, IIE_mu, including the cluster-average and individual-average versions).
IDE, IIE_M1, IIE_M2, and IIE_mu are, respectively, the interventional direct effect, indirect effect via M_1 (the first mediator in "Mnames"), indirect effect via M_2 (the second mediator in "Mnames"), and indirect effect due to the mediators' mutual dependence.
}
\description{
Estimating Causal Mediation Effects in Multiple-Mediator Analyses with Clustered Data
}
